@@ -29,7 +29,7 @@
     "framer-motion": "^6.2.9",
     "openai": "^4.0.0",
     "query-string": "^8.2.0",
-    "react": "^17.0.2",
+    "react": "^18",
     "react-apexcharts": "^1.4.0",
     "react-awesome-button": "^6.5.1",
     "react-beautiful-dnd": "^13.1.1",
@@ -37,7 +37,8 @@
     "react-circular-progressbar": "^2.0.4",
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18",
+    "react-error-boundary": "4.0.0",
     "react-hot-toast": "^2.4.1",
     "react-loader-spinner": "^5.1.2",
     "react-player": "^2.16.0",
@@ -46,7 +47,7 @@
     "react-scripts": "^3.0.1",
     "react-split": "^2.0.14",
     "react-tabs": "^3.2.3",
-    "react-toastify": "^9.0.4",
+    "react-toastify": "^11.0.2",
     "styled-components": "^6.1.13",
     "timeago.js": "^4.0.0-beta.3",
     "web-vitals": "^2.1.0"
